ISerializers: int to long since truly RdId type is long

+++ rd-net/RdFramework/Impl/Serializers.cs
@@ -219,7 +219,7 @@ namespace JetBrains.Rd.Impl
       Register(ReadEnum<T>, WriteEnum<T>);
     }
 
-    public void Register<T>(CtxReadDelegate<T> reader, CtxWriteDelegate<T> writer, int? predefinedId = null)
+    public void Register<T>(CtxReadDelegate<T> reader, CtxWriteDelegate<T> writer, long? predefinedId = null)
     {
       #if !NET35
       if (!myBackgroundRegistrar.IsInsideProcessing)
